test: de-duplicate useClusterIP upstream verification into shared helper
Extract the identical ClusterIP upstream-server verification used by the NginxProxy and UpstreamSettingsPolicy functional tests into expectUpstreamToUseClusterIP in system_suite_test.go, and call it from both tests.
