+++ tests/suite/nginxproxy_test.go
@@ -1,16 +1,13 @@
 package main
 
 import (
-	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/nginx/nginx-gateway-fabric/v2/tests/framework"
@@ -90,82 +87,8 @@ var _ = Describe("NginxProxy UseClusterIP", Ordered, Label("functional", "nginxp
 	})
 
 	Context("nginx config", func() {
-		var clusterIP string
-
-		BeforeAll(func() {
-			ctx, cancel := context.WithTimeout(context.Background(), timeoutConfig.GetStatusTimeout)
-			defer cancel()
-
-			var svc core.Service
-			err := resourceManager.Get(
-				ctx,
-				types.NamespacedName{Name: "coffee", Namespace: namespace},
-				&svc,
-			)
-			Expect(err).ToNot(HaveOccurred())
-			clusterIP = svc.Spec.ClusterIP
-		})
-
 		It("uses the Service ClusterIP as the upstream server", func() {
-			var serverAddr string
-			if strings.Contains(clusterIP, ":") {
-				serverAddr = fmt.Sprintf("[%s]:80", clusterIP)
-			} else {
-				serverAddr = fmt.Sprintf("%s:80", clusterIP)
-			}
-
-			upstreamName := fmt.Sprintf("%s_coffee_80", namespace)
-
-			if *plusEnabled {
-				// In NGINX Plus, upstream servers are managed via the Plus API and
-				// persisted in a state file rather than as server directives in the
-				// config. Read the state file to verify the ClusterIP is used.
-				Eventually(func() error {
-					ctx, cancel := context.WithTimeout(
-						context.Background(),
-						timeoutConfig.RequestTimeout,
-					)
-					defer cancel()
-
-					stateFileContent, err := resourceManager.GetNginxStateFile(
-						ctx, nginxPodName, namespace, upstreamName,
-					)
-					if err != nil {
-						return err
-					}
-
-					if !strings.Contains(stateFileContent, serverAddr) {
-						return fmt.Errorf(
-							"expected state file for upstream %s to contain server %s, got: %s",
-							upstreamName,
-							serverAddr,
-							stateFileContent,
-						)
-					}
-
-					return nil
-				}).
-					WithTimeout(timeoutConfig.GetStatusTimeout).
-					WithPolling(500 * time.Millisecond).
-					Should(Succeed())
-			} else {
-				Eventually(func() error {
-					conf, err := resourceManager.GetNginxConfig(nginxPodName, namespace, "")
-					if err != nil {
-						return err
-					}
-
-					return framework.ValidateNginxFieldExists(conf, framework.ExpectedNginxField{
-						Directive: "server",
-						Value:     serverAddr,
-						Upstream:  upstreamName,
-						File:      "http.conf",
-					})
-				}).
-					WithTimeout(timeoutConfig.GetStatusTimeout).
-					WithPolling(500 * time.Millisecond).
-					Should(Succeed())
-			}
+			expectUpstreamToUseClusterIP(nginxPodName, namespace, "coffee", fmt.Sprintf("%s_coffee_80", namespace))
 		})
 	})
 })
